Refactor to separate popovers to new module
